[breadboard-web] Convert Tour Guide Writer to new syntax.
