+++ seeds/breadboard-web/public/local-boards.json
@@ -103,12 +103,7 @@
   },
   {
     "title": "Tour Guide Writer 3",
-    "url": "/graphs/tour-guide-writer-3.json",
-    "version": "0.0.3"
-  },
-  {
-    "title": "Tour Guide Writer",
     "url": "/graphs/tour-guide-writer.json",
-    "version": "0.0.2"
+    "version": "0.0.3"
   }
 ]
